+++ src/boot/provider/hover-provider.ts
@@ -7,12 +7,14 @@ import * as vscode from 'vscode';
 import { Types } from "./types";
 import { AllKeywords } from '../jass/keyword';
 import { Options } from './options';
-import data, { DataGetter } from "./data";
+import data, { ConsumerMarkCode, DataGetter } from "./data";
 import { compare } from '../tool';
 import { convertPosition } from './tool';
 import { Func, Global, Library, Local, Member, Method, Take, Native, Struct, TextMacroDefine, Declaration, Type } from '../jass/ast';
 import { getTypeDesc } from './type-desc';
 import { tokenize } from '../jass/tokens';
+import { Document, lexically } from '../check/mark';
+import { MarkCodes } from '../war/mark';
 
 type Decl = Native|Func|Method|Library|Struct|Member|Global|Local|TextMacroDefine|Type;
 
@@ -291,6 +293,108 @@ class HoverProvider implements vscode.HoverProvider {
 
 }
 
+class MarkHoverProvider implements vscode.HoverProvider {
+
+  // 规定标识符长度
+  private _maxLength = 526;
+
+  private isNumber = function (val: string) {
+    var regPos = /^\d+(\.\d+)?$/; //非负浮点数
+    var regNeg = /^(-(([0-9]+\.[0-9]*[1-9][0-9]*)|([0-9]*[1-9][0-9]*\.[0-9]+)|([0-9]*[1-9][0-9]*)))$/; //负浮点数
+    if (regPos.test(val) || regNeg.test(val)) {
+      return true;
+    } else {
+      return false;
+    }
+  }
+
+  provideHover(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): vscode.ProviderResult<vscode.Hover> {
+
+    const key = document.getText(document.getWordRangeAtPosition(position));
+
+    if (key.length > this._maxLength) {
+      return null;
+    }
+
+    if (this.isNumber(key)) {
+      return null;
+    }
+
+    if (AllKeywords.includes(key)) {
+      return null;
+    }
+
+    // const type = Types.find(type => type === key);
+    // if (type) {
+    //   const markdownString = new vscode.MarkdownString().appendCodeblock(type);
+    //   markdownString.appendText("\n");
+    //   markdownString.appendText(getTypeDesc(type));
+    //   return new vscode.Hover(markdownString);
+    // }
+    console.info(key);
+
+    const fsPath = document.uri.fsPath;
+    // parseContent(fsPath, document.getText());
+
+    const hovers: vscode.MarkdownString[] = [];
+
+    if (Options.isSupportMark) {
+
+      const mark = lexically(new Document(document.uri.fsPath, document.lineAt(position.line).text)).find(mark => {
+        return mark.isMark() && mark.loc.start.position <= position.character && mark.loc.end.position >= position.character;
+      });
+      
+
+      
+      if (mark) {
+        const markValue = mark.value();
+        
+        const targetMark = MarkCodes.find(markCode => `'${markCode.code}'` == markValue)
+
+        if (targetMark) {
+          const ms = new vscode.MarkdownString();
+          ms.appendMarkdown(`***${targetMark.name}***`);
+          ms.appendText("\n");
+          ms.appendMarkdown(targetMark.tip);
+          ms.appendText("\n");
+          ms.appendCodeblock(`'${targetMark.code}'`);
+
+          ms.appendMarkdown("***@type***(" + (targetMark.type ? targetMark.type : "未知") + ")")
+          ms.appendMarkdown("  \n")
+          ms.appendMarkdown("***@race***(" + (targetMark.race ? targetMark.race : "未知") + ")")
+          ms.appendMarkdown("  \n")
+          ms.appendMarkdown("***@kind***(" + (targetMark.kind ? targetMark.kind : "未知") + ")");
+          hovers.push(ms);
+        }
+
+        const comsumerTargetMark = ConsumerMarkCode.instance(document.uri).getPresets().find(preset => `'${preset.code}'` == markValue);
+       
+
+        if (comsumerTargetMark) {
+          const ms = new vscode.MarkdownString();
+          ms.appendMarkdown(`***${comsumerTargetMark.name}***`);
+          ms.appendText("\n");
+          ms.appendMarkdown(comsumerTargetMark.descript);
+          ms.appendText("\n");
+          ms.appendCodeblock(`'${comsumerTargetMark.code}'`);
+
+          ms.appendMarkdown("***@type***(" + (comsumerTargetMark.type ? comsumerTargetMark.type : "未知") + ")")
+          ms.appendMarkdown("  \n")
+          ms.appendMarkdown("***@race***(" + (comsumerTargetMark.race ? comsumerTargetMark.race : "未知") + ")")
+          ms.appendMarkdown("  \n")
+          ms.appendMarkdown("***@kind***(" + (comsumerTargetMark.kind ? comsumerTargetMark.kind : "未知") + ")");
+          hovers.push(ms);
+        }
+      }
+
+
+    }
+
+    return new vscode.Hover([...hovers]);
+  }
+
+}
+
 vscode.languages.registerHoverProvider("jass", new HoverProvider());
 
-
+vscode.languages.registerHoverProvider("jass", new MarkHoverProvider());